pyproject.toml
[project]
name = "mailficker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
